Simple Options Menu Added
A simple options menu has been created in HTML buttons next to the
canvas to use. These consist of changing whether the mouse will drag or
change the tiles, place the grid in the center of the canvas. There is
also a way to select the type of tile to place. These join the existing
shrink and expand grid buttons.